Split code for Scala 2 and 3
